testing out some ideas for parsing the classes
Not entirely sure if TOML is the best kind of language for representing
this. I fear that YAML might be more suited for this *shudders*.

We also might just step away from the human readable handles.
Programatically it's much easier to work with the straight numbers and
the range is large enough that some logic can be put into it.