finish inital design
So, the base code is in order now. I'll probably put this on master
and a `0.0.1` release after a review. When the config is changed it
detects that there is a difference between the config tree and
system tree.

Some things that need to be fleshed out before another release:
* filters
* more qdisc implementation
* more class implementations

I'm curious to see as to how well the current config stucture will map
onto the other pieces of qdiscs and classes.